# Obsevatoire_de_recherche/src/Backend/laravel10/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "This is the API documentation for my application.",
        "contact": {
            "email": "your-email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API server"
        }
    ],
    "paths": {
...
        "/api/ressources/projets": {
            "get": {
                "tags": [
                    "Projets"
                ],
                "summary": "Get list of all projets",
                "operationId": "a3b932cfd57c73136eb4b2fa01a42572",
                "responses": {
                    "200": {
                        "description": "A list of projets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TblProjet"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Projets"
                ],
                "summary": "Create a new projet",
                "operationId": "6e8ca8aa36c1ec13c4099fdde3b8a1f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TblProjet"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Projet created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TblProjet"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/ressources/projets/{id}": {
            "get": {
                "tags": [
                    "Projets"
                ],
                "summary": "Get a projet by ID",
                "operationId": "56967419d0c2502e6033bae10a4869a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Projet details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TblProjet"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Projet not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Projets"
                ],
                "summary": "Update a projet",
                "operationId": "ec65b48be4c10005366c2f4b632caa7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TblProjet"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Projet updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TblProjet"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Projet not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Projets"
                ],
                "summary": "Delete a projet",
                "operationId": "c369f0f51a91ce45b2f980037e9d19a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Projet deleted"
                    },
                    "404": {
                        "description": "Projet not found"
                    }
                }
            }
        },
...
        "/api/ressources/universites": {
            "post": {
                "tags": [
                    "Universites"
                ],
                "summary": "Create a new universite",
                "operationId": "ce7c46bac6fd3ab66247c6429fceab99",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TblUniversite"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Universite created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TblUniversite"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "api/ressources/universites/{id}": {
            "get": {
                "tags": [
                    "Universites"
                ],
                "summary": "Get a universite by ID",
                "operationId": "a5ea75eece62ff814219080183c65960",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Universite details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TblUniversite"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Universite not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Universites"
                ],
                "summary": "Update a universite",
                "operationId": "5990c5f5a5ff292ebf532b1624bd0cba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TblUniversite"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Universite updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TblUniversite"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Universite not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Universites"
                ],
                "summary": "Delete a universite",
                "operationId": "d78dc031101a407f37bb7816dd50b982",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Universite deleted"
                    },
                    "404": {
                        "description": "Universite not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TblProjet": {
                "title": "TblProjet",
                "required": [
                    "titre_projet",
                    "descript_projet",
                    "tbl_niveau_id",
                    "user_id",
                    "tbl_categorie_id"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the projet",
                        "type": "integer",
                        "format": "int64"
                    },
                    "titre_projet": {
                        "description": "Title of the projet",
                        "type": "string"
                    },
                    "descript_projet": {
                        "description": "Description of the projet",
                        "type": "string"
                    },
                    "tbl_niveau_id": {
                        "description": "The ID of the associated niveau",
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "description": "The ID of the associated user",
                        "type": "integer",
                        "format": "int64"
                    },
                    "tbl_categorie_id": {
                        "description": "The ID of the associated categorie",
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "type": "object"
            },
...
            "TblUniversite": {
                "title": "TblUniversite",
                "required": [
                    "nom_univ",
                    "localite_univ",
                    "email_univ",
                    "boite_postale"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the universite",
                        "type": "integer",
                        "format": "int64"
                    },
                    "nom_univ": {
                        "description": "Name of the universite",
                        "type": "string"
                    },
                    "localite_univ": {
                        "description": "Location of the universite",
                        "type": "string"
                    },
                    "email_univ": {
                        "description": "Email of the universite",
                        "type": "string"
                    },
                    "boite_postale": {
                        "description": "Postal box of the universite",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
...
}